feat!(egui): move `AnySection` into it's own file, improve spacing, size and child section styling in section
