+++ egui/v1/src/widgets/settings/mod.rs
@@ -1,7 +1,6 @@
-mod renderer;
-
 mod settings;
 pub use settings::*;
 
 pub mod button;
 pub mod section;
+pub mod any_section;
